+++ Magic/CHANGELOG.md
@@ -20,6 +20,7 @@
  - Fix `CheckInventory` action `blocked_enchantments` (fixes mars air supply behavior)
  - Fix `always_on: true` magic blocks not starting back up on server restart
  - Fix wands dealing melee damage even if not on the melee list (this includes the lightsaber hilt)
+ - Fix players not having permission to use arena signs by default
 
 ## Changes